gratuitous demonstration of extending interfaces
It's also interesting that I did not have to import ModuleResolutionResult
Even before I renamed the file to .d.ts
This is because when it did not participate in the module system, the interface appeared
in the global space. It was part of that mysterious module definition loading
that I am trying to figure out. It was ambient ...
I wonder if I can get it to tell me about ambient symbols? or namespaces? that aren't modules?
... but getAmbientModules only shows the ones from @types/